Use upstream ScheduledSearchController as the base for scheduled alerts.
