+++ module/FinnaConsole/src/FinnaConsole/Service/AbstractService.php
@@ -27,9 +27,6 @@
  */
 namespace FinnaConsole\Service;
 
-use Zend\Log\Logger;
-use Zend\Log\Writer\Stream;
-
 /**
  * Abstract base class for console services.
  *
@@ -41,195 +38,6 @@ use Zend\Log\Writer\Stream;
  */
 abstract class AbstractService implements ConsoleServiceInterface
 {
-    /**
-     * Symbolic link name for institution default view
-     */
-    const DEFAULT_PATH = 'default';
-
-    /**
-     * Logger
-     *
-     * @var Logger
-     */
-    protected $logger = null;
-
-    /**
-     * Error logger
-     *
-     * @var Logger
-     */
-    protected $errLogger = null;
-
-    /**
-     * Init logging.
-     *
-     * @return void
-     */
-    public function initLogging()
-    {
-        $writer = new Stream('php://output');
-        $this->logger = new Logger();
-        $this->logger->addWriter($writer);
-
-        $writer = new Stream('php://stderr');
-        $this->errLogger = new Logger();
-        $this->errLogger->addWriter($writer);
-    }
-
-    /**
-     * Log an exception triggered by ZF2 for administrative purposes.
-     *
-     * @param \Exception $error Exception to log
-     *
-     * @return void
-     */
-    public function logException($error)
-    {
-        // We need to build a variety of pieces so we can supply
-        // information at five different verbosity levels:
-        $baseError = get_class($error) . ' : ' . $error->getMessage();
-        $prev = $error->getPrevious();
-        while ($prev) {
-            $baseError .= ' ; ' . get_class($prev) . ' : ' . $prev->getMessage();
-            $prev = $prev->getPrevious();
-        }
-        $backtrace = "\nBacktrace:\n";
-        if (is_array($error->getTrace())) {
-            foreach ($error->getTrace() as $line) {
-                if (!isset($line['file'])) {
-                    $line['file'] = 'unlisted file';
-                }
-                if (!isset($line['line'])) {
-                    $line['line'] = 'unlisted';
-                }
-                $backtraceLine = $line['file'] .
-                    ' line ' . $line['line'] . ' - ' .
-                    (isset($line['class']) ? 'class = ' . $line['class'] . ', ' : '')
-                    . 'function = ' . $line['function'];
-                $backtrace .= "{$backtraceLine}\n";
-                if (!empty($line['args'])) {
-                    $args = [];
-                    foreach ($line['args'] as $i => $arg) {
-                        $args[] = $i . ' = ' . $this->argumentToString($arg);
-                    }
-                    $backtraceLine .= ', args: ' . implode(', ', $args);
-                } else {
-                    $backtraceLine .= ', args: none.';
-                }
-                $backtrace .= "{$backtraceLine}\n";
-            }
-        }
-
-        $this->logger->err($baseError . $backtrace);
-    }
-
-    /**
-     * Convert function argument to a loggable string
-     *
-     * @param mixed $arg Argument
-     *
-     * @return string
-     */
-    protected function argumentToString($arg)
-    {
-        if (is_object($arg)) {
-            return get_class($arg) . ' Object';
-        }
-        if (is_array($arg)) {
-            $args = [];
-            foreach ($arg as $key => $item) {
-                $args[] = "$key => " . $this->argumentToString($item);
-            }
-            return 'array(' . implode(', ', $args) . ')';
-        }
-        if (is_bool($arg)) {
-            return $arg ? 'true' : 'false';
-        }
-        if (is_int($arg) || is_float($arg)) {
-            return (string)$arg;
-        }
-        if (null === $arg) {
-            return 'null';
-        }
-        return "'$arg'";
-    }
-
-    /**
-     * Output a message with a timestamp
-     *
-     * @param string $msg Message
-     *
-     * @return void
-     */
-    protected function msg($msg)
-    {
-        $msg = '[' . getmypid() . "] $msg";
-        $this->logger->info($msg);
-    }
-
-    /**
-     * Output an error message with a timestamp
-     *
-     * @param string $msg          Message
-     * @param string $publishedMsg Published version of the error message. Must
-     * exclude any privacy-related information. Use '=' for same as $msg.
-     *
-     * @return void
-     */
-    protected function err($msg, $publishedMsg = '')
-    {
-        if ($publishedMsg) {
-            if ('=' === $publishedMsg) {
-                $publishedMsg = $msg;
-            }
-            $publishedMsg = '[' . getmypid() . "] $publishedMsg";
-            $this->errLogger->err($publishedMsg);
-        }
-        $msg = '[' . getmypid() . "] $msg";
-        $this->logger->err($msg);
-    }
-
-    /**
-     * Output a warning message with a timestamp
-     *
-     * @param string $msg Message
-     *
-     * @return void
-     */
-    protected function warn($msg)
-    {
-        $msg = '[' . getmypid() . "] $msg";
-        $this->logger->warn($msg);
-    }
-
-    /**
-     * Resolve path to the view directory.
-     *
-     * @param string $institution Institution
-     * @param string $view        View
-     *
-     * @return string|boolean view path or false on error
-     */
-    protected function resolveViewPath($institution, $view = false)
-    {
-        if (!$view) {
-            $view = $this::DEFAULT_PATH;
-            if (isset($this->datasourceConfig[$institution]['mainView'])) {
-                list($institution, $view)
-                    = explode(
-                        '/',
-                        $this->datasourceConfig[$institution]['mainView'], 2
-                    );
-            }
-        }
-        $path = "{$this->viewBaseDir}/$institution/$view";
-
-        // Assume that view is functional if index.php exists.
-        if (!is_file("$path/public/index.php")) {
-            $this->err("Could not resolve view path for $institution/$view", '=');
-            return false;
-        }
-
-        return $path;
-    }
+    use ConsoleLoggerTrait;
+    use ViewPathTrait;
 }
